massa_sc_runtime

Function run_function

pub fn run_function(
    interface: &dyn Interface,
    rt_module: RuntimeModule,
    function: &str,
    param: &[u8],
    limit: u64,
    gas_costs: GasCosts,
    condom_limits: CondomLimits,
) -> Result<Response, VMError>

Library Input, take a module wasm built with the massa environment, run a function of that module with the given parameter:

import { print } from "massa-sc-std";

export function hello_world(_args: string): i32 {
    print("hello world");
    return 0;
}