Function massa_sc_runtime::run_main

pub fn run_main(
    interface: &dyn Interface,
    rt_module: RuntimeModule,
    limit: u64,
    gas_costs: GasCosts,
    max_exports: Option<usize>
) -> Result<Response, VMError>

Library Input, take a module wasm built with the massa environment, must have a main function inside written in AssemblyScript:

import { print } from "massa-sc-std";

export function main(_args: string): i32 {
    print("hello world");
    return 0;
}

Return: the remaining gas.