Crate massa_sc_runtime

Structs

CondomLimits

GasCosts

Response

That’s what is returned when a module is executed correctly since the end

Enums

Compiler

Enum listing the available compilers

RuntimeModule

VMError

Traits

Interface

InterfaceClone

Functions

run_function

Library Input, take a module wasm built with the massa environment, run a function of that module with the given parameter:

run_main

Library Input, take a module wasm built with the massa environment, must have a main function inside written in AssemblyScript: