Trait massa_sc_runtime::Interface

pub trait Interface: Send + Sync + InterfaceClone {

    fn init_call(&self, address: &str, raw_coins: u64) -> Result<Vec<u8>> { ... }
    fn finish_call(&self) -> Result<()> { ... }
    fn get_module(&self, address: &str) -> Result<Vec<u8>> { ... }
    fn get_balance(&self) -> Result<u64> { ... }
    fn get_balance_for(&self, address: &str) -> Result<u64> { ... }
    fn transfer_coins(&self, to_address: &str, raw_amount: u64) -> Result<()> { ... }
    fn transfer_coins_for(
        &self,
        from_address: &str,
        to_address: &str,
        raw_amount: u64
    ) -> Result<()> { ... }
    fn get_call_coins(&self) -> Result<u64> { ... }
    fn raw_set_bytecode_for(&self, address: &str, bytecode: &[u8]) -> Result<()> { ... }
    fn raw_set_bytecode(&self, bytecode: &[u8]) -> Result<()> { ... }
    fn create_module(&self, module: &[u8]) -> Result<String> { ... }
    fn print(&self, message: &str) -> Result<()> { ... }
    fn raw_get_data(&self, key: &str) -> Result<Vec<u8>> { ... }
    fn raw_set_data(&self, key: &str, value: &[u8]) -> Result<()> { ... }
    fn raw_append_data(&self, key: &str, value: &[u8]) -> Result<()> { ... }
    fn raw_delete_data(&self, key: &str) -> Result<()> { ... }
    fn raw_get_data_for(&self, address: &str, key: &str) -> Result<Vec<u8>> { ... }
    fn raw_set_data_for(
        &self,
        address: &str,
        key: &str,
        value: &[u8]
    ) -> Result<()> { ... }
    fn raw_append_data_for(
        &self,
        address: &str,
        key: &str,
        value: &[u8]
    ) -> Result<()> { ... }
    fn raw_delete_data_for(&self, address: &str, key: &str) -> Result<()> { ... }
    fn has_data(&self, key: &str) -> Result<bool> { ... }
    fn has_data_for(&self, address: &str, _key: &str) -> Result<bool> { ... }
    fn get_op_keys(&self) -> Result<Vec<Vec<u8>>> { ... }
    fn has_op_key(&self, key: &[u8]) -> Result<bool> { ... }
    fn get_op_data(&self, key: &[u8]) -> Result<Vec<u8>> { ... }
    fn hash(&self, data: &[u8]) -> Result<String> { ... }
    fn signature_verify(
        &self,
        data: &[u8],
        signature: &str,
        public_key: &str
    ) -> Result<bool> { ... }
    fn address_from_public_key(&self, public_key: &str) -> Result<String> { ... }
    fn get_time(&self) -> Result<u64> { ... }
    fn unsafe_random(&self) -> Result<i64> { ... }
    fn get_current_period(&self) -> Result<u64> { ... }
    fn get_current_thread(&self) -> Result<u8> { ... }
    fn module_called(&self) -> Result<()> { ... }
    fn exit_success(&self) -> Result<()> { ... }
    fn get_owned_addresses(&self) -> Result<Vec<String>> { ... }
    fn get_call_stack(&self) -> Result<Vec<String>> { ... }
    fn generate_event(&self, _event: String) -> Result<()> { ... }
    fn send_message(
        &self,
        target_address: &str,
        target_handler: &str,
        validity_start: (u64, u8),
        validity_end: (u64, u8),
        max_gas: u64,
        gas_price: u64,
        raw_coins: u64,
        data: &[u8]
    ) -> Result<()> { ... }

}

Provided Methods

fn init_call(&self, address: &str, raw_coins: u64) -> Result<Vec<u8>>

Prepare the execution of a module at the given address and transfer a given amount of coins

fn finish_call(&self) -> Result<()>

Finish a call

fn get_module(&self, address: &str) -> Result<Vec<u8>>

Requires the module in the given address

fn get_balance(&self) -> Result<u64>

Get the SCE ledger balance for the current address. Defaults to zero if the address is not found.

fn get_balance_for(&self, address: &str) -> Result<u64>

Get the SCE ledger balance for an address. Defaults to zero if the address is not found.

fn transfer_coins(&self, to_address: &str, raw_amount: u64) -> Result<()>

Transfer an amount from the address on the current call stack to a target address.

fn transfer_coins_for(
&self,
from_address: &str,
to_address: &str,
raw_amount: u64
) -> Result<()>

Transfer an amount from the specified address to a target address.

fn get_call_coins(&self) -> Result<u64>

Get the amount of coins that have been made available for use by the caller of the currently executing code.

fn raw_set_bytecode_for(&self, address: &str, bytecode: &[u8]) -> Result<()>

Sets the executable bytecode at a target address. The target address must exist and the current context must have access rights.

fn raw_set_bytecode(&self, bytecode: &[u8]) -> Result<()>

Sets the executable bytecode at a current address.

fn create_module(&self, module: &[u8]) -> Result<String>

Requires a new address that contains the sent &u8

fn print(&self, message: &str) -> Result<()>

Print function for examples

fn raw_get_data(&self, key: &str) -> Result<Vec<u8>>

fn raw_set_data(&self, key: &str, value: &[u8]) -> Result<()>

fn raw_append_data(&self, key: &str, value: &[u8]) -> Result<()>

fn raw_delete_data(&self, key: &str) -> Result<()>

fn raw_get_data_for(&self, address: &str, key: &str) -> Result<Vec<u8>>

Requires the data at the address

fn raw_set_data_for(&self, address: &str, key: &str, value: &[u8]) -> Result<()>

fn raw_append_data_for(&self, address: &str, key: &str, value: &[u8]) -> Result<()>

fn raw_delete_data_for(&self, address: &str, key: &str) -> Result<()>

fn has_data(&self, key: &str) -> Result<bool>

Requires to replace the data in the current address

Note: The execution lib will always use the current context address for the update

fn has_data_for(&self, address: &str, _key: &str) -> Result<bool>

fn get_op_keys(&self) -> Result<Vec<Vec<u8>>>

Return operation datastore keys

fn has_op_key(&self, key: &[u8]) -> Result<bool>

Check if key is in operation datastore

fn get_op_data(&self, key: &[u8]) -> Result<Vec<u8>>

Return operation datastore keys

fn hash(&self, data: &[u8]) -> Result<String>

fn signature_verify(
&self,
data: &[u8],
signature: &str,
public_key: &str
) -> Result<bool>

fn address_from_public_key(&self, public_key: &str) -> Result<String>

fn get_time(&self) -> Result<u64>

Returns the current time (millisecond unix timestamp)

fn unsafe_random(&self) -> Result<i64>

Returns a random number (unsafe: can be predicted and manipulated)

fn get_current_period(&self) -> Result<u64>

Returns the period of the current execution slot

fn get_current_thread(&self) -> Result<u8>

Returns the thread of the current execution slot

fn module_called(&self) -> Result<()>

fn exit_success(&self) -> Result<()>

fn get_owned_addresses(&self) -> Result<Vec<String>>

Expect to return a list of owned addresses

Required on smart-contract execute the imported function assembly_script_get_owned_addresses

fn get_call_stack(&self) -> Result<Vec<String>>

Expect to return a list of addresses in the call stack

Required on smart-contract execute the imported function assembly_script_get_call_stack

fn generate_event(&self, _event: String) -> Result<()>

fn send_message(
&self,
target_address: &str,
target_handler: &str,
validity_start: (u64, u8),
validity_end: (u64, u8),
max_gas: u64,
gas_price: u64,
raw_coins: u64,
data: &[u8]
) -> Result<()>

Sends an async message

Arguments

target_address - Destination address hash in format string

target_handler - Name of the message handling function

validity_start - Tuple containing the period and thread of the validity start slot

validity_end - Tuple containing the period and thread of the validity end slot

max_gas - Maximum gas for the message execution

gas_price - Price of one gas unit

coins - Coins of the sender

data - Message data

Implementations

impl dyn Interface

pub fn get_data<T: DeserializeOwned>(&self, key: &str) -> Result<T>

pub fn set_data<T: Serialize>(&self, key: &str, value: &T) -> Result<()>

pub fn get_data_for<T: DeserializeOwned>(
&self,
address: &str,
key: &str
) -> Result<T>

pub fn set_data_for<T: Serialize>(
&self,
address: &str,
key: &str,
value: &T
) -> Result<()>