Trait massa_sc_runtime::Interface

pub trait Interface: Send + Sync + InterfaceClone {

    // Required methods
    fn init_call(&self, address: &str, raw_coins: u64) -> Result<Vec<u8>>;
    fn init_call_wasmv1(
        &self,
        address: &str,
        raw_coins: NativeAmount
    ) -> Result<Vec<u8>>;
    fn finish_call(&self) -> Result<()>;
    fn get_balance(&self) -> Result<u64>;
    fn get_balance_for(&self, address: &str) -> Result<u64>;
    fn get_balance_wasmv1(
        &self,
        address: Option<String>
    ) -> Result<NativeAmount>;
    fn transfer_coins(&self, to_address: &str, raw_amount: u64) -> Result<()>;
    fn transfer_coins_for(
        &self,
        from_address: &str,
        to_address: &str,
        raw_amount: u64
    ) -> Result<()>;
    fn transfer_coins_wasmv1(
        &self,
        to_address: String,
        raw_amount: NativeAmount,
        from_address: Option<String>
    ) -> Result<()>;
    fn get_call_coins_wasmv1(&self) -> Result<NativeAmount>;
    fn raw_set_bytecode(&self, bytecode: &[u8]) -> Result<()>;
    fn raw_set_bytecode_for(&self, address: &str, bytecode: &[u8]) -> Result<()>;
    fn set_bytecode_wasmv1(
        &self,
        bytecode: &[u8],
        address: Option<String>
    ) -> Result<()>;
    fn create_module(&self, module: &[u8]) -> Result<String>;
    fn print(&self, message: &str) -> Result<()>;
    fn get_keys(&self, prefix: Option<&[u8]>) -> Result<BTreeSet<Vec<u8>>>;
    fn get_keys_for(
        &self,
        address: &str,
        prefix: Option<&[u8]>
    ) -> Result<BTreeSet<Vec<u8>>>;
    fn get_ds_keys_wasmv1(
        &self,
        prefix: &[u8],
        address: Option<String>
    ) -> Result<BTreeSet<Vec<u8>>>;
    fn raw_get_data(&self, key: &[u8]) -> Result<Vec<u8>>;
    fn raw_get_data_for(&self, address: &str, key: &[u8]) -> Result<Vec<u8>>;
    fn get_ds_value_wasmv1(
        &self,
        key: &[u8],
        address: Option<String>
    ) -> Result<Vec<u8>>;
    fn raw_set_data(&self, key: &[u8], value: &[u8]) -> Result<()>;
    fn raw_set_data_for(
        &self,
        address: &str,
        key: &[u8],
        value: &[u8]
    ) -> Result<()>;
    fn set_ds_value_wasmv1(
        &self,
        key: &[u8],
        value: &[u8],
        address: Option<String>
    ) -> Result<()>;
    fn raw_append_data(&self, key: &[u8], value: &[u8]) -> Result<()>;
    fn raw_append_data_for(
        &self,
        address: &str,
        key: &[u8],
        value: &[u8]
    ) -> Result<()>;
    fn append_ds_value_wasmv1(
        &self,
        key: &[u8],
        value: &[u8],
        address: Option<String>
    ) -> Result<()>;
    fn raw_delete_data(&self, key: &[u8]) -> Result<()>;
    fn raw_delete_data_for(&self, address: &str, key: &[u8]) -> Result<()>;
    fn delete_ds_entry_wasmv1(
        &self,
        key: &[u8],
        address: Option<String>
    ) -> Result<()>;
    fn has_data(&self, key: &[u8]) -> Result<bool>;
    fn has_data_for(&self, address: &str, key: &[u8]) -> Result<bool>;
    fn ds_entry_exists_wasmv1(
        &self,
        key: &[u8],
        address: Option<String>
    ) -> Result<bool>;
    fn raw_get_bytecode(&self) -> Result<Vec<u8>>;
    fn raw_get_bytecode_for(&self, address: &str) -> Result<Vec<u8>>;
    fn get_bytecode_wasmv1(&self, address: Option<String>) -> Result<Vec<u8>>;
    fn get_op_keys(&self, prefix: Option<&[u8]>) -> Result<Vec<Vec<u8>>>;
    fn get_op_keys_wasmv1(&self, prefix: &[u8]) -> Result<Vec<Vec<u8>>>;
    fn op_entry_exists(&self, key: &[u8]) -> Result<bool>;
    fn get_op_data(&self, key: &[u8]) -> Result<Vec<u8>>;
    fn caller_has_write_access(&self) -> Result<bool>;
    fn hash(&self, data: &[u8]) -> Result<[u8; 32]>;
    fn hash_blake3(&self, bytes: &[u8]) -> Result<[u8; 32]>;
    fn signature_verify(
        &self,
        data: &[u8],
        signature: &str,
        public_key: &str
    ) -> Result<bool>;
    fn evm_signature_verify(
        &self,
        message: &[u8],
        signature: &[u8],
        public_key: &[u8]
    ) -> Result<bool>;
    fn evm_get_address_from_pubkey(&self, public_key: &[u8]) -> Result<Vec<u8>>;
    fn evm_get_pubkey_from_signature(
        &self,
        hash: &[u8],
        signature: &[u8]
    ) -> Result<Vec<u8>>;
    fn is_address_eoa(&self, address: &str) -> Result<bool>;
    fn address_from_public_key(&self, public_key: &str) -> Result<String>;
    fn validate_address(&self, address: &str) -> Result<bool>;
    fn get_time(&self) -> Result<u64>;
    fn unsafe_random(&self) -> Result<i64>;
    fn unsafe_random_f64(&self) -> Result<f64>;
    fn unsafe_random_wasmv1(&self, num_bytes: u64) -> Result<Vec<u8>>;
    fn get_current_period(&self) -> Result<u64>;
    fn get_current_thread(&self) -> Result<u8>;
    fn get_current_slot(&self) -> Result<Slot>;
    fn get_owned_addresses(&self) -> Result<Vec<String>>;
    fn get_call_stack(&self) -> Result<Vec<String>>;
    fn generate_event(&self, _event: String) -> Result<()>;
    fn generate_event_wasmv1(&self, _event: Vec<u8>) -> Result<()>;
    fn get_module(
        &self,
        bytecode: &[u8],
        limit: u64
    ) -> Result<(RuntimeModule, u64)>;
    fn get_tmp_module(
        &self,
        bytecode: &[u8],
        limit: u64
    ) -> Result<(RuntimeModule, u64)>;
    fn send_message(
        &self,
        target_address: &str,
        target_handler: &str,
        validity_start: (u64, u8),
        validity_end: (u64, u8),
        max_gas: u64,
        raw_fee: u64,
        raw_coins: u64,
        data: &[u8],
        filter: Option<(&str, Option<&[u8]>)>
    ) -> Result<()>;
    fn get_origin_operation_id(&self) -> Result<Option<String>>;
    fn hash_sha256(&self, bytes: &[u8]) -> Result<[u8; 32]>;
    fn hash_keccak256(&self, bytes: &[u8]) -> Result<[u8; 32]>;
    fn native_amount_from_str_wasmv1(
        &self,
        amount: &str
    ) -> Result<NativeAmount>;
    fn native_amount_to_string_wasmv1(
        &self,
        amount: &NativeAmount
    ) -> Result<String>;
    fn check_native_amount_wasmv1(&self, amount: &NativeAmount) -> Result<bool>;
    fn add_native_amount_wasmv1(
        &self,
        amount1: &NativeAmount,
        amount2: &NativeAmount
    ) -> Result<NativeAmount>;
    fn sub_native_amount_wasmv1(
        &self,
        amount1: &NativeAmount,
        amount2: &NativeAmount
    ) -> Result<NativeAmount>;
    fn scalar_mul_native_amount_wasmv1(
        &self,
        amount: &NativeAmount,
        factor: u64
    ) -> Result<NativeAmount>;
    fn scalar_div_rem_native_amount_wasmv1(
        &self,
        dividend: &NativeAmount,
        divisor: u64
    ) -> Result<(NativeAmount, NativeAmount)>;
    fn div_rem_native_amount_wasmv1(
        &self,
        dividend: &NativeAmount,
        divisor: &NativeAmount
    ) -> Result<(u64, NativeAmount)>;
    fn check_address_wasmv1(&self, to_check: &str) -> Result<bool>;
    fn check_pubkey_wasmv1(&self, to_check: &str) -> Result<bool>;
    fn check_signature_wasmv1(&self, to_check: &str) -> Result<bool>;
    fn get_address_category_wasmv1(
        &self,
        to_check: &str
    ) -> Result<AddressCategory>;
    fn get_address_version_wasmv1(&self, address: &str) -> Result<u64>;
    fn get_pubkey_version_wasmv1(&self, pubkey: &str) -> Result<u64>;
    fn get_signature_version_wasmv1(&self, signature: &str) -> Result<u64>;
    fn checked_add_native_time_wasmv1(
        &self,
        time1: &NativeTime,
        time2: &NativeTime
    ) -> Result<NativeTime>;
    fn checked_sub_native_time_wasmv1(
        &self,
        time1: &NativeTime,
        time2: &NativeTime
    ) -> Result<NativeTime>;
    fn checked_mul_native_time_wasmv1(
        &self,
        time: &NativeTime,
        factor: u64
    ) -> Result<NativeTime>;
    fn checked_scalar_div_native_time_wasmv1(
        &self,
        dividend: &NativeTime,
        divisor: u64
    ) -> Result<(NativeTime, NativeTime)>;
    fn checked_div_native_time_wasmv1(
        &self,
        dividend: &NativeTime,
        divisor: &NativeTime
    ) -> Result<(u64, NativeTime)>;
    fn base58_check_to_bytes_wasmv1(&self, s: &str) -> Result<Vec<u8>>;
    fn bytes_to_base58_check_wasmv1(&self, bytes: &[u8]) -> String;
    fn compare_address_wasmv1(
        &self,
        left: &str,
        right: &str
    ) -> Result<ComparisonResult>;
    fn compare_native_amount_wasmv1(
        &self,
        left: &NativeAmount,
        right: &NativeAmount
    ) -> Result<ComparisonResult>;
    fn compare_native_time_wasmv1(
        &self,
        left: &NativeTime,
        right: &NativeTime
    ) -> Result<ComparisonResult>;
    fn compare_pub_key_wasmv1(
        &self,
        left: &str,
        right: &str
    ) -> Result<ComparisonResult>;

    // Provided method
    fn get_call_coins(&self) -> Result<u64> { ... }

}

Required Methods

fn init_call(&self, address: &str, raw_coins: u64) -> Result<Vec<u8>>

Prepare the execution of a module at the given address and transfer a given amount of coins

fn init_call_wasmv1( &self, address: &str, raw_coins: NativeAmount ) -> Result<Vec<u8>>

Prepare the execution of a module at the given address and transfer a given amount of coins

fn finish_call(&self) -> Result<()>

Finish a call

fn get_balance(&self) -> Result<u64>

Get the SCE ledger balance for the current address. Defaults to zero if the address is not found.

fn get_balance_for(&self, address: &str) -> Result<u64>

Get the SCE ledger balance for an address. Defaults to zero if the address is not found.

fn get_balance_wasmv1(&self, address: Option<String>) -> Result<NativeAmount>

fn transfer_coins(&self, to_address: &str, raw_amount: u64) -> Result<()>

Transfer an amount from the address on the current call stack to a target address.

fn transfer_coins_for( &self, from_address: &str, to_address: &str, raw_amount: u64 ) -> Result<()>

Transfer an amount from the specified address to a target address.

fn transfer_coins_wasmv1( &self, to_address: String, raw_amount: NativeAmount, from_address: Option<String> ) -> Result<()>

fn get_call_coins_wasmv1(&self) -> Result<NativeAmount>

Get the native amount of coins that have been made available for use by the caller of the currently executing code.

fn raw_set_bytecode(&self, bytecode: &[u8]) -> Result<()>

Sets the executable bytecode at a current address.

fn raw_set_bytecode_for(&self, address: &str, bytecode: &[u8]) -> Result<()>

Sets the executable bytecode at a target address. The target address must exist and the current context must have access rights.

fn set_bytecode_wasmv1( &self, bytecode: &[u8], address: Option<String> ) -> Result<()>

fn create_module(&self, module: &[u8]) -> Result<String>

Requires a new address that contains the sent &u8

fn print(&self, message: &str) -> Result<()>

Print function for examples

fn get_keys(&self, prefix: Option<&[u8]>) -> Result<BTreeSet<Vec<u8>>>

Return datastore keys Will only return keys with a given prefix if provided in args

fn get_keys_for( &self, address: &str, prefix: Option<&[u8]> ) -> Result<BTreeSet<Vec<u8>>>

Return datastore keys Will only return keys with a given prefix if provided in args

fn get_ds_keys_wasmv1( &self, prefix: &[u8], address: Option<String> ) -> Result<BTreeSet<Vec<u8>>>

fn raw_get_data(&self, key: &[u8]) -> Result<Vec<u8>>

Return the datastore value of the corresponding key

fn raw_get_data_for(&self, address: &str, key: &[u8]) -> Result<Vec<u8>>

Requires the data at the address

fn get_ds_value_wasmv1( &self, key: &[u8], address: Option<String> ) -> Result<Vec<u8>>

fn raw_set_data(&self, key: &[u8], value: &[u8]) -> Result<()>

Set the datastore value for the corresponding key

fn raw_set_data_for( &self, address: &str, key: &[u8], value: &[u8] ) -> Result<()>

Set the datastore value for the corresponding key of the given address

fn set_ds_value_wasmv1( &self, key: &[u8], value: &[u8], address: Option<String> ) -> Result<()>

fn raw_append_data(&self, key: &[u8], value: &[u8]) -> Result<()>

Append a value to the current datastore value for the corresponding key

fn raw_append_data_for( &self, address: &str, key: &[u8], value: &[u8] ) -> Result<()>

Append a value to the current datastore value for the corresponding key and the given address

fn append_ds_value_wasmv1( &self, key: &[u8], value: &[u8], address: Option<String> ) -> Result<()>

fn raw_delete_data(&self, key: &[u8]) -> Result<()>

Delete a datastore entry

fn raw_delete_data_for(&self, address: &str, key: &[u8]) -> Result<()>

Delete a datastore entry at of the given address

fn delete_ds_entry_wasmv1( &self, key: &[u8], address: Option<String> ) -> Result<()>

fn has_data(&self, key: &[u8]) -> Result<bool>

Requires to replace the data in the current address

Note: The execution lib will always use the current context address for the update

fn has_data_for(&self, address: &str, key: &[u8]) -> Result<bool>

Check if a datastore entry exists

fn ds_entry_exists_wasmv1( &self, key: &[u8], address: Option<String> ) -> Result<bool>

fn raw_get_bytecode(&self) -> Result<Vec<u8>>

Returns bytecode of the current address

fn raw_get_bytecode_for(&self, address: &str) -> Result<Vec<u8>>

Returns bytecode of the target address

fn get_bytecode_wasmv1(&self, address: Option<String>) -> Result<Vec<u8>>

fn get_op_keys(&self, prefix: Option<&[u8]>) -> Result<Vec<Vec<u8>>>

Return operation datastore keys

fn get_op_keys_wasmv1(&self, prefix: &[u8]) -> Result<Vec<Vec<u8>>>

fn op_entry_exists(&self, key: &[u8]) -> Result<bool>

Check if operation in datastore exists

fn get_op_data(&self, key: &[u8]) -> Result<Vec<u8>>

Return operation datastore data for a given key

fn caller_has_write_access(&self) -> Result<bool>

Check whether or not the caller has write access in the current context

fn hash(&self, data: &[u8]) -> Result<[u8; 32]>

Hash data

fn hash_blake3(&self, bytes: &[u8]) -> Result<[u8; 32]>

Returns the blake3 hash of the given bytes

fn signature_verify( &self, data: &[u8], signature: &str, public_key: &str ) -> Result<bool>

Verify signature

fn evm_signature_verify( &self, message: &[u8], signature: &[u8], public_key: &[u8] ) -> Result<bool>

Verify signature (EVM)

fn evm_get_address_from_pubkey(&self, public_key: &[u8]) -> Result<Vec<u8>>

Get address from public key (EVM)

fn evm_get_pubkey_from_signature( &self, hash: &[u8], signature: &[u8] ) -> Result<Vec<u8>>

Get public key from signature (EVM)

fn is_address_eoa(&self, address: &str) -> Result<bool>

Return true if the address is a User address, false if it is an SC address

fn address_from_public_key(&self, public_key: &str) -> Result<String>

Convert a public key to an address

fn validate_address(&self, address: &str) -> Result<bool>

Validate an address

fn get_time(&self) -> Result<u64>

Returns the current time (millisecond unix timestamp)

fn unsafe_random(&self) -> Result<i64>

Returns a random number (unsafe: can be predicted and manipulated)

fn unsafe_random_f64(&self) -> Result<f64>

Returns a random number (unsafe: can be predicted and manipulated)

fn unsafe_random_wasmv1(&self, num_bytes: u64) -> Result<Vec<u8>>

Returns a random number (unsafe: can be predicted and manipulated)

fn get_current_period(&self) -> Result<u64>

Returns the period of the current execution slot

fn get_current_thread(&self) -> Result<u8>

Returns the thread of the current execution slot

fn get_current_slot(&self) -> Result<Slot>

Returns the current execution slot

fn get_owned_addresses(&self) -> Result<Vec<String>>

Expect to return a list of owned addresses

Required on smart-contract execute the imported function assembly_script_get_owned_addresses

fn get_call_stack(&self) -> Result<Vec<String>>

Expect to return a list of addresses in the call stack

Required on smart-contract execute the imported function assembly_script_get_call_stack

fn generate_event(&self, _event: String) -> Result<()>

Generate a smart contract event

fn generate_event_wasmv1(&self, _event: Vec<u8>) -> Result<()>

Generate a smart contract event

fn get_module( &self, bytecode: &[u8], limit: u64 ) -> Result<(RuntimeModule, u64)>

For the given bytecode:

Get the corresponding runtime module if it already exists

Compile it if not

fn get_tmp_module( &self, bytecode: &[u8], limit: u64 ) -> Result<(RuntimeModule, u64)>

Same as get_module without cache

fn send_message( &self, target_address: &str, target_handler: &str, validity_start: (u64, u8), validity_end: (u64, u8), max_gas: u64, raw_fee: u64, raw_coins: u64, data: &[u8], filter: Option<(&str, Option<&[u8]>)> ) -> Result<()>

Sends an async message

Arguments

target_address - Destination address hash in format string

target_handler - Name of the message handling function

validity_start - Tuple containing the period and thread of the validity start slot

validity_end - Tuple containing the period and thread of the validity end slot

max_gas - Maximum gas for the message execution

raw_fee - Fee to be paid for message execution

coins - Coins of the sender

data - Message data

fn get_origin_operation_id(&self) -> Result<Option<String>>

fn hash_sha256(&self, bytes: &[u8]) -> Result<[u8; 32]>

fn hash_keccak256(&self, bytes: &[u8]) -> Result<[u8; 32]>

fn native_amount_from_str_wasmv1(&self, amount: &str) -> Result<NativeAmount>

fn native_amount_to_string_wasmv1( &self, amount: &NativeAmount ) -> Result<String>

fn check_native_amount_wasmv1(&self, amount: &NativeAmount) -> Result<bool>

fn add_native_amount_wasmv1( &self, amount1: &NativeAmount, amount2: &NativeAmount ) -> Result<NativeAmount>

fn sub_native_amount_wasmv1( &self, amount1: &NativeAmount, amount2: &NativeAmount ) -> Result<NativeAmount>

fn scalar_mul_native_amount_wasmv1( &self, amount: &NativeAmount, factor: u64 ) -> Result<NativeAmount>

fn scalar_div_rem_native_amount_wasmv1( &self, dividend: &NativeAmount, divisor: u64 ) -> Result<(NativeAmount, NativeAmount)>

fn div_rem_native_amount_wasmv1( &self, dividend: &NativeAmount, divisor: &NativeAmount ) -> Result<(u64, NativeAmount)>

fn check_address_wasmv1(&self, to_check: &str) -> Result<bool>

fn check_pubkey_wasmv1(&self, to_check: &str) -> Result<bool>

fn check_signature_wasmv1(&self, to_check: &str) -> Result<bool>

fn get_address_category_wasmv1(&self, to_check: &str) -> Result<AddressCategory>

fn get_address_version_wasmv1(&self, address: &str) -> Result<u64>

fn get_pubkey_version_wasmv1(&self, pubkey: &str) -> Result<u64>

fn get_signature_version_wasmv1(&self, signature: &str) -> Result<u64>

fn checked_add_native_time_wasmv1( &self, time1: &NativeTime, time2: &NativeTime ) -> Result<NativeTime>

fn checked_sub_native_time_wasmv1( &self, time1: &NativeTime, time2: &NativeTime ) -> Result<NativeTime>

fn checked_mul_native_time_wasmv1( &self, time: &NativeTime, factor: u64 ) -> Result<NativeTime>

fn checked_scalar_div_native_time_wasmv1( &self, dividend: &NativeTime, divisor: u64 ) -> Result<(NativeTime, NativeTime)>

fn checked_div_native_time_wasmv1( &self, dividend: &NativeTime, divisor: &NativeTime ) -> Result<(u64, NativeTime)>

fn base58_check_to_bytes_wasmv1(&self, s: &str) -> Result<Vec<u8>>

fn bytes_to_base58_check_wasmv1(&self, bytes: &[u8]) -> String

fn compare_address_wasmv1( &self, left: &str, right: &str ) -> Result<ComparisonResult>

fn compare_native_amount_wasmv1( &self, left: &NativeAmount, right: &NativeAmount ) -> Result<ComparisonResult>

fn compare_native_time_wasmv1( &self, left: &NativeTime, right: &NativeTime ) -> Result<ComparisonResult>

fn compare_pub_key_wasmv1( &self, left: &str, right: &str ) -> Result<ComparisonResult>

Provided Methods

fn get_call_coins(&self) -> Result<u64>

Get the amount of coins that have been made available for use by the caller of the currently executing code.

Implementations

impl dyn Interface

pub fn get_data<T: DeserializeOwned>(&self, key: &[u8]) -> Result<T>

pub fn get_data_for<T: DeserializeOwned>( &self, address: &str, key: &[u8] ) -> Result<T>

pub fn set_data<T: Serialize>(&self, key: &[u8], value: &T) -> Result<()>

pub fn set_data_for<T: Serialize>( &self, address: &str, key: &[u8], value: &T ) -> Result<()>