Trait massa_sc_runtime::Interface

pub trait Interface: Send + Sync + InterfaceClone {

    fn init_call(&self, _address: &Address, _raw_coins: u64) -> Result<Bytecode> { ... }

    fn finish_call(&self) -> Result<()> { ... }

    fn get_balance(&self) -> Result<u64> { ... }

    fn get_balance_for(&self, _address: &Address) -> Result<u64> { ... }

    fn transfer_coins(
        &self, 
        _to_address: &Address, 
        _raw_amount: u64
    ) -> Result<()> { ... }

    fn transfer_coins_for(
        &self, 
        _from_address: &Address, 
        _to_address: &Address, 
        _raw_amount: u64
    ) -> Result<()> { ... }

    fn get_call_coins(&self) -> Result<u64> { ... }

    fn update_module(
        &self, 
        _address: &Address, 
        _module: &Bytecode
    ) -> Result<()> { ... }

    fn create_module(&self, _module: &Bytecode) -> Result<Address> { ... }

    fn get_data_for(
        &self, 
        _address: &Address, 
        _key: &MassaHash
    ) -> Result<Vec<u8>> { ... }

    fn print(&self, _message: &str) -> Result<()> { ... }

    fn set_data_for(
        &self, 
        _address: &Address, 
        _key: &MassaHash, 
        _value: &[u8]
    ) -> Result<()> { ... }

    fn get_data(&self, _key: &str) -> Result<Vec<u8>> { ... }

    fn set_data(&self, _key: &str, _value: &[u8]) -> Result<()> { ... }

    fn has_data(&self, _key: &str) -> Result<bool> { ... }

    fn has_data_for(&self, _address: &Address, _key: &str) -> Result<bool> { ... }

    fn hash(&self, _data: &[u8]) -> Result<MassaHash> { ... }

    fn signature_verify(
        &self, 
        _data: &[u8], 
        _signature: &Signature, 
        _public_key: &PublicKey
    ) -> Result<bool> { ... }

    fn address_from_public_key(
        &self, 
        _public_key: &PublicKey
    ) -> Result<Address> { ... }

    fn get_time(&self) -> Result<u64> { ... }

    fn unsafe_random(&self) -> Result<i64> { ... }

    fn module_called(&self) -> Result<()> { ... }

    fn get_owned_addresses(&self) -> Result<Vec<Address>> { ... }

    fn get_call_stack(&self) -> Result<Vec<Address>> { ... }

    fn generate_event(&self, _event: String) -> Result<()> { ... }

}

Provided methods

fn init_call(&self, _address: &Address, _raw_coins: u64) -> Result<Bytecode>

Prepare the execution of a module at the given address and transfer a given amount of coins

fn finish_call(&self) -> Result<()>

Finish a call

fn get_balance(&self) -> Result<u64>

Get the SCE ledger balance for the crrent address. Defaults to zero if the address is not found.

fn get_balance_for(&self, _address: &Address) -> Result<u64>

Get the SCE ledger balance for an address. Defaults to zero if the address is not found.

fn transfer_coins(&self, _to_address: &Address, _raw_amount: u64) -> Result<()>

Transfer an amount from the address on the current call stack to a target address.

fn transfer_coins_for(
&self,
_from_address: &Address,
_to_address: &Address,
_raw_amount: u64
) -> Result<()>

Transfer an amount from the specified address to a target address.

fn get_call_coins(&self) -> Result<u64>

Get the amount of coins that have been made available for use by the caller of the currently executing code.

fn update_module(&self, _address: &Address, _module: &Bytecode) -> Result<()>

Requires to replace the module at the current address

Note: The execution lib will use the current context address for the update module and the new bytecode

fn create_module(&self, _module: &Bytecode) -> Result<Address>

Requires a new address that contains the bytecode sended

fn get_data_for(&self, _address: &Address, _key: &MassaHash) -> Result<Vec<u8>>

Requires the data at the address

fn print(&self, _message: &str) -> Result<()>

Print function for examples

fn set_data_for(
&self,
_address: &Address,
_key: &MassaHash,
_value: &[u8]
) -> Result<()>

Requires to replace the data in the current address

Note: The execution lib will allways use the current context address for the update

fn get_data(&self, _key: &str) -> Result<Vec<u8>>

fn set_data(&self, _key: &str, _value: &[u8]) -> Result<()>

fn has_data(&self, _key: &str) -> Result<bool>

fn has_data_for(&self, _address: &Address, _key: &str) -> Result<bool>

fn hash(&self, _data: &[u8]) -> Result<MassaHash>

fn signature_verify(
&self,
_data: &[u8],
_signature: &Signature,
_public_key: &PublicKey
) -> Result<bool>

fn address_from_public_key(&self, _public_key: &PublicKey) -> Result<Address>

fn get_time(&self) -> Result<u64>

Returns the current time (millisecond unix timestamp)

fn unsafe_random(&self) -> Result<i64>

Returns a random number (unsafe: can be predicted and manipulated)

fn module_called(&self) -> Result<()>

fn get_owned_addresses(&self) -> Result<Vec<Address>>

fn get_call_stack(&self) -> Result<Vec<Address>>

fn generate_event(&self, _event: String) -> Result<()>